requirements.txt
requests==2.26.0
bs4==0.0.1
pytz==2021.1
colorama==0.4.4
python-dotenv==0.19.2